% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{add_ground_truth_counts}
\alias{add_ground_truth_counts}
\title{Add ground truth counts to nowcasts}
\usage{
add_ground_truth_counts(nowcasts, ground_truth_nowcasts)
}
\arguments{
\item{nowcasts}{Data.frame of nowcasts (any method) without initial/final
counts}

\item{ground_truth_nowcasts}{Data.frame of nowcasts from a reference method
(e.g., baselinenowcast) that contains initial_count and final_count columns}
}
\value{
nowcasts with initial_count and final_count joined from the ground
truth
}
\description{
Add ground truth counts to nowcasts
}
