% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weekly_data.R
\name{clean_data}
\alias{clean_data}
\title{Clean data and add attributes like season and other indicators}
\usage{
clean_data(
  weekly_data,
  max_delay,
  nowcast_date_range,
  prev_season_date_range,
  delay_est_date_range
)
}
\arguments{
\item{weekly_data}{Data.frame of cases by epiweek}

\item{max_delay}{Integer indicating max delay in weeks.}

\item{nowcast_date_range}{Range of dates to be nowcasted.}

\item{prev_season_date_range}{Range of dates for previous season.}

\item{delay_est_date_range}{Range of dates used for historical delay
estimation}
}
\value{
Cleaned dataframe filtered by max delay with season indicated.
}
\description{
Clean data and add attributes like season and other indicators
}
