% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{clean_dph_nowcasts_ag}
\alias{clean_dph_nowcasts_ag}
\title{Clean DPH nowcasts by age group}
\usage{
clean_dph_nowcasts_ag(ma_nowcasts, eval_horizon)
}
\arguments{
\item{ma_nowcasts}{Raw nowcasts from DPH}
}
\value{
only the MA nowcasts with only the columns required
}
\description{
Clean DPH nowcasts by age group
}
