% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{clean_madph_nowcasts}
\alias{clean_madph_nowcasts}
\title{Clean MADPH nowcasts}
\usage{
clean_madph_nowcasts(ma_nowcasts, eval_horizon)
}
\arguments{
\item{ma_nowcasts}{Raw nowcasts from MADPH}
}
\value{
only the MA nowcasts with only the columns required
}
\description{
Clean MADPH nowcasts
}
