% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bnc.R
\name{fit_bnc_age_groups}
\alias{fit_bnc_age_groups}
\title{Fit the baselinenowcast method to age-groups}
\usage{
fit_bnc_age_groups(
  all_data,
  nowcast_date,
  pathogen_i,
  model,
  eval_horizon,
  max_delay,
  quantiles_for_scoring,
  scale_factor = 3,
  prop_delay = 0.5,
  draws = 1000
)
}
\arguments{
\item{all_data}{Data.frame of incident cases by reference date and report
date by week for multiple age groups and pathogens}

\item{nowcast_date}{Date to produce the nowcast for.}

\item{pathogen_i}{Pathogen to nowcast.}

\item{eval_horizon}{Number of weeks to evaluation and save the nowcast.}

\item{max_delay}{Maximum delay in weeks.}

\item{quantiles_for_scoring}{Vector of quantiles to score.}

\item{scale_factor}{Scale factor on maximum delay of the amount of data to
be used to train the baselinenowcast model.}

\item{prop_delay}{Proportion of all training volume to use for delay
estimation}

\item{draws}{Number of draws to save}
}
\value{
Quantiled dataframe of nowcasts with initial and final case counts
alongside it.
}
\description{
Fit the baselinenowcast method to age-groups
}
