% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_group_scores_fig.R
\name{get_bar_chart_by_ag}
\alias{get_bar_chart_by_ag}
\title{Get a plot of decomposed WIS by age group for each model permutation}
\usage{
get_bar_chart_by_ag(
  scores,
  pathogen,
  fig_file_name = NULL,
  fig_file_dir = file.path("output", "figs", "supp")
)
}
\arguments{
\item{scores}{Dataframe of the scores by age group
and model}

\item{pathogen}{Character string indicating pathogen to plot}

\item{fig_file_name}{name of figure}

\item{fig_file_dir}{filepath to save figure}
}
\value{
ggplot object
}
\description{
Get a plot of decomposed WIS by age group for each model permutation
}
