% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_scores_fig.R
\name{get_bar_chart_coverage}
\alias{get_bar_chart_coverage}
\title{Bar chart of 95th percent coverage for each pathogen and model}
\usage{
get_bar_chart_coverage(
  coverage,
  title = NULL,
  fig_file_name = NULL,
  fig_file_dir = file.path("output", "figs", "supp")
)
}
\arguments{
\item{coverage}{Dataframe of 95\% interval coverage by model and pathogen}

\item{title}{Title of plot}

\item{fig_file_name}{Character string indicating name of fig}

\item{fig_file_dir}{Character string indicating the filepathir}
}
\value{
ggplot object
}
\description{
Bar chart of 95th percent coverage for each pathogen and model
}
