% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_comparison_figs.R
\name{get_bar_chart_mean_delay_comp}
\alias{get_bar_chart_mean_delay_comp}
\title{Get a bar chart of the mean delay for each method and by season}
\usage{
get_bar_chart_mean_delay_comp(ma_delay, delay_dfs_bnc)
}
\arguments{
\item{ma_delay}{Dataframe of MA delay distribution}

\item{delay_dfs_bnc}{Dataframe of delay distribution from latest data}
}
\value{
ggplot of the mean delay by method and season
}
\description{
Get a bar chart of the mean delay for each method and by season
}
