% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_volume_opt_figs.R
\name{get_bar_chart_tv_scores}
\alias{get_bar_chart_tv_scores}
\title{Get bar charts of WIS breakdown by pathogen and scale factor/prop_delay}
\usage{
get_bar_chart_tv_scores(scores_su)
}
\arguments{
\item{scores_su}{Data.frame of scores by pathogen, nowcast date, model,
and reference date}
}
\value{
ggplot object
}
\description{
Get bar charts of WIS breakdown by pathogen and scale factor/prop_delay
}
