% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_fig.R
\name{get_cases_by_season_plot}
\alias{get_cases_by_season_plot}
\title{Cases over time by age group and pathogen plot}
\usage{
get_cases_by_season_plot(weekly_data, season_to_plot = NULL)
}
\arguments{
\item{weekly_data}{Dataframe of cases by reference date, report date, week,
and pathogen}

\item{season_to_plot}{Character string indicating what season to plot,
default is NULL which will include all dates in the data}

\item{by_age_group}{Boolean indicating whether or not to plot cases by age
group, default is TRUE.}
}
\value{
ggplot
}
\description{
Cases over time by age group and pathogen plot
}
