% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_fig.R
\name{get_cases_plot}
\alias{get_cases_plot}
\title{Cases over time by age group and pathogen plot}
\usage{
get_cases_plot(weekly_data, season_to_plot = NULL, by_age_group = TRUE)
}
\arguments{
\item{weekly_data}{Dataframe of cases by reference date, report date, week,
and pathogen}

\item{season_to_plot}{Character string indicating what season to plot,
default is NULL which will include all dates in the data}
}
\value{
ggplot
}
\description{
Cases over time by age group and pathogen plot
}
