% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weekly_data.R
\name{get_daily_data}
\alias{get_daily_data}
\title{Get properly formatted daily data with delays}
\usage{
get_daily_data(raw_data)
}
\arguments{
\item{raw_data}{}
}
\value{
formatted daily data
}
\description{
Get properly formatted daily data with delays
}
