% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_fig.R
\name{get_delay_cdf_plot}
\alias{get_delay_cdf_plot}
\title{Get plot of cdf}
\usage{
get_delay_cdf_plot(weekly_data)
}
\arguments{
\item{weekly_data}{Dataframe of cases by reference date, report date, week,
and pathogen}
}
\value{
ggplot
}
\description{
Get plot of cdf
}
