% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_comparison_figs.R
\name{get_delay_df}
\alias{get_delay_df}
\title{Get delay distribution from data using baselinenowcast}
\usage{
get_delay_df(data, nowcast_date, pathogen, max_delay)
}
\arguments{
\item{data}{Dataframe of weekly data by age group and pathogen}

\item{nowcast_date}{Date of the nowcast}

\item{pathogen}{Character string indicating the pathogen}

\item{max_delay}{Integer indicating the maximum delay}
}
\value{
Data.frame of delay distribution pdf and cdf for each nowcast date
and pathogen
}
\description{
Get delay distribution from data using baselinenowcast
}
