% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_fig.R
\name{get_delay_t_by_season}
\alias{get_delay_t_by_season}
\title{Delay over time by age group and pathogen and season}
\usage{
get_delay_t_by_season(weekly_data, season_to_plot = NULL, ylims = TRUE)
}
\arguments{
\item{weekly_data}{Dataframe of cases by reference date, report date, week,
and pathogen}

\item{season_to_plot}{Character string indicating what season to plot,
default is NULL which will include all dates in the data}

\item{y_lims}{Boolean indicating to set the ylim values. Default is TRUE.}
}
\value{
ggplot
}
\description{
Delay over time by age group and pathogen and season
}
