% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{get_dph_nowcasts}
\alias{get_dph_nowcasts}
\title{Get DPH nowcasts}
\usage{
get_dph_nowcasts(fp)
}
\arguments{
\item{fp}{filepath of DPH nowcasts}
}
\value{
Data.frame of nowcasts from DPH model
}
\description{
Get DPH nowcasts
}
