% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_comparison_figs.R
\name{get_ma_delay_data}
\alias{get_ma_delay_data}
\title{Get MA delay data from local multipliers files}
\usage{
get_ma_delay_data(fp_prefix, pathogen, max_delay)
}
\arguments{
\item{fp_prefix}{Filepath prefix}

\item{pathogen}{Character string indicating pathogen}

\item{max_delay}{Integer indicating the maximum delay}
}
\value{
Data.frame with median and 95\% PI of PDF and CDF
}
\description{
Get MA delay data from local multipliers files
}
