% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_comparison_figs.R
\name{get_ma_multipliers_from_file}
\alias{get_ma_multipliers_from_file}
\title{Get MA multipliers from file}
\usage{
get_ma_multipliers_from_file(fp_prefix, pathogen, max_delay)
}
\arguments{
\item{fp_prefix}{Filepath prefix}

\item{pathogen}{Character string indicating pathogen}

\item{max_delay}{Integer indicating the maximum delay}
}
\value{
Data.frame with multipliers
}
\description{
Get MA multipliers from file
}
