% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{get_madph_nowcasts}
\alias{get_madph_nowcasts}
\title{Get MADPH nowcasts}
\usage{
get_madph_nowcasts(fp)
}
\arguments{
\item{fp}{filepath of MADPH nowcasts}
}
\value{
Data.frame of nowcasts from MADPH model
}
\description{
Get MADPH nowcasts
}
