% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_scores_fig.R
\name{get_overall_scores}
\alias{get_overall_scores}
\title{Get bar charts of WIS breakdown by pathogen and model}
\usage{
get_overall_scores(scores_su)
}
\arguments{
\item{scores_su}{Data.frame of scores by pathogen, nowcast date, model,
and reference date}
}
\value{
ggplot object
}
\description{
Get bar charts of WIS breakdown by pathogen and model
}
