% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_group_scores_fig.R
\name{get_plot_ag_nowcasts_vs_data}
\alias{get_plot_ag_nowcasts_vs_data}
\title{Get a plot illustrating nowcasts at certain dates for each age group for
a specific pathogen}
\usage{
get_plot_ag_nowcasts_vs_data(
  nowcasts,
  all_data,
  max_delay,
  pathogen_i,
  nowcast_dates_to_plot = NULL
)
}
\arguments{
\item{nowcasts}{Dataframe of the combined quantiles across
horizons and nowcast dates}

\item{max_delay}{Integer indicating the maximum delay, used to create the
evaluation data.}

\item{nowcast_dates_to_plot}{Vector of character strings of the dates you
wish to plot, default is \code{NULL} which will plot all of them}
}
\value{
ggplot object
}
\description{
Get a plot illustrating nowcasts at certain dates for each age group for
a specific pathogen
}
