% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_fig.R
\name{get_plot_delay_by_season}
\alias{get_plot_delay_by_season}
\title{Get a plot of the delay overall by season}
\usage{
get_plot_delay_by_season(
  weekly_data,
  fig_file_name = NULL,
  fig_file_dir = file.path("output", "figs", "supp")
)
}
\arguments{
\item{weekly_data}{data at weekly scale by reference and report date}

\item{fig_file_name}{Character string indicating name of fig}

\item{fig_file_dir}{Character string indicating the filepath}
}
\value{
ggplot object
}
\description{
Get a plot of the delay overall by season
}
