% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_comparison_figs.R
\name{get_plot_delay_comparison}
\alias{get_plot_delay_comparison}
\title{Get a bar chart comparing the mean delays}
\usage{
get_plot_delay_comparison(ma_delay, delay_dfs_bnc, plot_type = "cdf")
}
\arguments{
\item{ma_delay}{Dataframe of delay distribution from MA data}

\item{delay_dfs_bnc}{Dataframe of delay distributions from data}

\item{plot_type}{Character string indicating whether to plot the pdf or
the cdf}
}
\value{
ggplot showing each delay distribution for both data and MA delay,
faceted by pathogen
}
\description{
Get a bar chart comparing the mean delays
}
