% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_fig.R
\name{get_plot_delay_distrib_seas}
\alias{get_plot_delay_distrib_seas}
\title{Get a plot of the delay distributions by season}
\usage{
get_plot_delay_distrib_seas(weekly_data)
}
\arguments{
\item{weekly_data}{data at weekly scale by reference and report date}
}
\value{
ggplot object
}
\description{
Get a plot of the delay distributions by season
}
