% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_comparison_figs.R
\name{get_plot_multipliers}
\alias{get_plot_multipliers}
\title{Get a plot comparing the multipliers}
\usage{
get_plot_multipliers(multipliers_combined)
}
\arguments{
\item{multipliers_combined}{Dataframe of multipliers from both methods}
}
\value{
ggplot of multipliers
}
\description{
Get a plot comparing the multipliers
}
