% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_fig.R
\name{get_plot_norm_delay_in_season}
\alias{get_plot_norm_delay_in_season}
\title{Get a plot of normalized within-season delay variability}
\usage{
get_plot_norm_delay_in_season(weekly_data)
}
\arguments{
\item{weekly_data}{data at weekly scale by reference and report date}
}
\value{
ggplot object
}
\description{
Get a plot of normalized within-season delay variability
}
