% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_group_scores_fig.R
\name{get_plot_nowcasts_over_time}
\alias{get_plot_nowcasts_over_time}
\title{Get a plot of the nowcasts over time by horizon and age group for a specific
pathogen}
\usage{
get_plot_nowcasts_over_time(
  nowcasts,
  horizon_to_plot,
  age_group_to_plot = "00-04",
  pathogen_to_plot = "bar",
  fig_file_name = NULL,
  fig_file_dir = file.path("output", "figs", "supp")
)
}
\arguments{
\item{nowcasts}{Age group specific nowcasts}

\item{horizon_to_plot}{Integer indicating horizon (in weeks) to plot}

\item{age_group_to_plot}{Character string indicating the age group to plot}

\item{pathogen_to_plot}{Character string indicating what pathogen to plot}

\item{fig_file_name}{Character string indicating name of the figure}

\item{fig_file_dir}{Filepath to save figure}
}
\value{
ggplot object faceted by model showing nowcasts for the chosen
horizon and age group
}
\description{
Get a plot of the nowcasts over time by horizon and age group for a specific
pathogen
}
