% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_fig.R
\name{get_plot_prop_visits_t}
\alias{get_plot_prop_visits_t}
\title{Prop visits reported over time by age group and pathogen plot}
\usage{
get_plot_prop_visits_t(weekly_data, season_to_plot = NULL)
}
\arguments{
\item{weekly_data}{Dataframe of cases by reference date, report date, week,
and pathogen}

\item{season_to_plot}{Character string indicating what season to plot,
default is NULL which will include all dates in the data}
}
\value{
ggplot
}
\description{
Prop visits reported over time by age group and pathogen plot
}
