% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_style.R
\name{get_plot_theme}
\alias{get_plot_theme}
\title{Get standardized plot theme to add to figures}
\usage{
get_plot_theme(dates = TRUE)
}
\arguments{
\item{dates}{Boolean indicating whether the x-axis represents dates.}
}
\value{
a theme object to add to a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object
}
\description{
Get standardized plot theme to add to figures
}
