% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_scores_fig.R
\name{get_state_wis_over_time_plot}
\alias{get_state_wis_over_time_plot}
\title{State-level WIS over time + final data, faceted by pathogen}
\usage{
get_state_wis_over_time_plot(
  all_data,
  scores,
  nowcasts,
  max_delay,
  fig_file_name = NULL,
  fig_file_dir = file.path("output", "figs", "supp")
)
}
\arguments{
\item{all_data}{Weekly data by reference and report date for all pathogens}

\item{scores}{Data.frame of scores by pathogen, nowcast date, model,
and reference date}

\item{nowcasts}{Dataframe of the combined quantiles across
horizons and nowcast dates}

\item{max_delay}{Integer indicating the maximum delay, used to create the
evaluation data.}

\item{fig_file_name}{Character string indicating name of fig}

\item{fig_file_dir}{Character string indicating the filepath}
}
\value{
patchwork fig
}
\description{
State-level WIS over time + final data, faceted by pathogen
}
