% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_volume_opt_figs.R
\name{get_table_min_wis}
\alias{get_table_min_wis}
\title{Get table of the minimum wis}
\usage{
get_table_min_wis(scores_su)
}
\arguments{
\item{scores_su}{Data.frame of scores by pathogen, nowcast date, model,
and reference date}
}
\description{
Get table of the minimum wis
}
