% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weekly_data.R
\name{get_weekly_data}
\alias{get_weekly_data}
\title{Get weekly data from raw data}
\usage{
get_weekly_data(raw_data)
}
\arguments{
\item{raw_data}{Data.frame of cases by reference and report date and
pathogen}
}
\value{
Case counts by epiweek
}
\description{
Get weekly data from raw data
}
