% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bnc.R
\name{implement_dph_method}
\alias{implement_dph_method}
\title{Implement the DPH method}
\usage{
implement_dph_method(
  multipliers,
  age_group,
  all_data,
  nowcast_date,
  pathogen_i,
  eval_horizon,
  max_delay,
  model_name
)
}
\arguments{
\item{multipliers}{DPH multipliers estimated from 2023 data}

\item{age_group}{Character string indicating age group to nowcast}

\item{all_data}{Clean weekly data for all age groups}

\item{nowcast_date}{Date of the nowcast}

\item{pathogen_i}{Character string indicating pathogen to nowcast}

\item{eval_horizon}{Integer indicating number of weeks to evaluate}

\item{max_delay}{Maximum delay}

\item{model_name}{Character string indicating name of the model}
}
\value{
Nowcast dataframe
}
\description{
Implement the DPH method
}
