% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_group_scores_fig.R
\name{make_ag_nowcast_comp_fig}
\alias{make_ag_nowcast_comp_fig}
\title{Make age group nowcast comparison figure}
\usage{
make_ag_nowcast_comp_fig(
  nowcasts_vs_data1,
  nowcasts_vs_data2,
  nowcasts_vs_data3,
  nowcasts_vs_data4,
  bar_chart_scores1,
  bar_chart_scores2,
  bar_chart_scores3,
  bar_chart_scores4,
  fig_file_name = NULL,
  fig_file_dir = file.path("output", "figs")
)
}
\arguments{
\item{nowcasts_vs_data1}{plot A}

\item{nowcasts_vs_data2}{plot B}

\item{nowcasts_vs_data3}{plot C}

\item{nowcasts_vs_data4}{plot D}

\item{bar_chart_scores1}{plot E}

\item{bar_chart_scores2}{plot F}

\item{bar_chart_scores3}{plot G}

\item{bar_chart_scores4}{plot H}

\item{fig_file_name}{name of figure}

\item{fig_file_dir}{filepath to save figure}
}
\value{
patchwork fig
}
\description{
Make age group nowcast comparison figure
}
