% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_fig.R
\name{make_comp_seasons_fig}
\alias{make_comp_seasons_fig}
\title{Make delay across multiple seasons}
\usage{
make_comp_seasons_fig(
  delay_over_time,
  case_counts,
  fig_file_name = NULL,
  fig_file_dir = file.path("output", "figs")
)
}
\arguments{
\item{delay_over_time}{plot delay over time}

\item{case_counts}{plot cases over time}

\item{fig_file_name}{Character string indicating name of fig}

\item{fig_file_dir}{Character string indicating the filepath}
}
\value{
patchwork fig
}
\description{
Make delay across multiple seasons
}
