% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_fig.R
\name{make_delay_fig}
\alias{make_delay_fig}
\title{Make delay figure}
\usage{
make_delay_fig(
  delay_over_time,
  case_counts,
  violin_plot_delay,
  season_to_plot,
  fig_file_name = NULL,
  fig_file_dir = file.path("output", "figs")
)
}
\arguments{
\item{delay_over_time}{plot delay over time}

\item{case_counts}{plot cases over time}

\item{violin_plot_delay}{plot mean delay distribution}

\item{season_to_plot}{Character string indicating the season}

\item{fig_file_name}{Character string indicating name of fig}

\item{fig_file_dir}{Character string indicating the filepath}
}
\value{
patchwork fig
}
\description{
Make delay figure
}
