% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_scores_fig.R
\name{make_state_nowcast_comp_fig}
\alias{make_state_nowcast_comp_fig}
\title{Make delay figure}
\usage{
make_state_nowcast_comp_fig(
  nowcasts_vs_data,
  bar_chart_scores1,
  bar_chart_scores2,
  bar_chart_scores3,
  bar_chart_scores4,
  fig_file_name = NULL,
  fig_file_dir = file.path("output", "figs")
)
}
\arguments{
\item{fig_file_name}{Character string indicating name of fig}

\item{fig_file_dir}{Character string indicating the filepath}

\item{delay_over_time}{plot delay over time}

\item{case_counts}{plot cases over time}

\item{violin_plot_delay}{plot mean delay distribution}

\item{season_to_plot}{Character string indicating the season}
}
\value{
patchwork fig
}
\description{
Make delay figure
}
