% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_fig.R
\name{make_within_between_season_fig}
\alias{make_within_between_season_fig}
\title{Make a two-panel comparison of within vs between season variability}
\usage{
make_within_between_season_fig(
  weekly_data,
  fig_file_name = NULL,
  fig_file_dir = file.path("output", "figs", "supp")
)
}
\arguments{
\item{weekly_data}{data at weekly scale by reference and report date}

\item{fig_file_name}{Character string indicating name of fig}

\item{fig_file_dir}{Character string indicating the filepath}
}
\value{
patchwork fig
}
\description{
Make a two-panel comparison of within vs between season variability
}
