% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quiet}
\alias{quiet}
\title{Suppress output and messages for code.}
\usage{
quiet(code)
}
\arguments{
\item{code}{Code to run quietly.}
}
\value{
The result of running the code.
}
\description{
Suppress output and messages for code.
}
\examples{
result <- quiet(message("This message should be suppressed"))
print(result)
}
