% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{read_pathogen_data}
\alias{read_pathogen_data}
\title{Load in data}
\usage{
read_pathogen_data(df, fp)
}
\arguments{
\item{df}{Grouped data.frame}

\item{fp}{File.path of data}
}
\description{
Load in data
}
