% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trajectories_to_quantiles}
\alias{trajectories_to_quantiles}
\title{Aggregate individual trajectory
timeseries or forecasts to quantile
timeseries or forecasts}
\usage{
trajectories_to_quantiles(
  trajectories,
  quantiles = c(0.01, 0.025, 1:19/20, 0.975, 0.99),
  timepoint_cols = "timepoint",
  value_col = "value",
  quantile_value_name = "quantile_value",
  quantile_level_name = "quantile_level",
  id_cols = NULL
)
}
\arguments{
\item{trajectories}{Tidy data frame or tibble
of trajectories}

\item{quantiles}{Quantiles to output for each
timepoint (default the FluSight/COVIDHub 2024-25 quantiles:
\code{c(0.01, 0.025, 1:19/20, 0.975, 0.99)}}

\item{timepoint_cols}{Name(s) of the column(s) in \code{trajectories}
that identifies unique timepoints. Default \code{"timepoint"}.}

\item{value_col}{name of the column in \code{trajectories}
with the trajectory values (for which we wish to
compute quantiles), e.g. \code{hosp}, \code{weekly_hosp}, \code{cases},
etc. Default \code{value}.}

\item{quantile_value_name}{What to name
the column containing quantile values in
the output table. Default \code{"quantile_value"}}

\item{quantile_level_name}{What to name
the column containing quantile levels in
the output table. Default \code{"quantile_level"}}

\item{id_cols}{additional id columns in
\code{trajectories} to group by before aggregating,
e.g. a \code{location} column if \code{trajectories} contains
trajectories over the same time period for multiple
locations, such as different US States and Territories.
If NULL, ignored. Default NULL.}
}
\description{
Given a tidy data frame of
trajectories, aggregate it to
a quantile timeseries for the
given quantile values.
}
\details{
From https://github.com/CDCgov/forecasttools/blob/main/R/trajectories_to_quantiles.R #nolint
}
