% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{clean_madph_nowcasts_ag}
\alias{clean_madph_nowcasts_ag}
\title{Clean MADPH nowcasts by age group}
\usage{
clean_madph_nowcasts_ag(ma_nowcasts)
}
\arguments{
\item{ma_nowcasts}{Raw nowcasts from MADPH}
}
\value{
only the MA nowcasts with only the columns required
}
\description{
Clean MADPH nowcasts by age group
}
