% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_fig.R
\name{get_mean_delay_over_time_plot}
\alias{get_mean_delay_over_time_plot}
\title{Delay over all across age groups}
\usage{
get_mean_delay_over_time_plot(
  weekly_data,
  ma_delay,
  fig_file_name = NULL,
  fig_file_dir = file.path("output", "figs", "supp")
)
}
\arguments{
\item{weekly_data}{Dataframe of cases by reference date, report date, week,
and pathogen}

\item{ma_delay}{Data.frame of delays from MADPH}

\item{fig_file_name}{Character string indicating name of fig}

\item{fig_file_dir}{Character string indicating the filepath}
}
\value{
ggplot
}
\description{
Delay over all across age groups
}
