% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bnc.R
\name{get_multipliers}
\alias{get_multipliers}
\title{Derive multipliers using MADPH methods but within this codebase}
\usage{
get_multipliers(all_data, source, this_age_group = "00+")
}
\arguments{
\item{all_data}{Dataframe of weekly cases by reference and report date
stratified by age group}

\item{source}{Character string indicating where data is from and its method}

\item{this_age_group}{Selected age group}
}
\value{
dataframe of median and 95\% CI for the pmf at each delay (in weeks)
}
\description{
Derive multipliers using MADPH methods but within this codebase
}
