% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bnc.R
\name{get_multipliers_from_daily_data_orig}
\alias{get_multipliers_from_daily_data_orig}
\title{Derive multipliers using MADPH methods but within this codebase, using
their original implementation}
\usage{
get_multipliers_from_daily_data_orig(
  all_data,
  max_delay,
  source,
  age_group = "00+"
)
}
\arguments{
\item{all_data}{Dataframe of daily cases by reference and report date
stratified by age group}

\item{max_delay}{Integer indicating maximum delay in weeks}

\item{source}{Character string indicating where data is from and its method}

\item{age_group}{Selected age group}
}
\value{
dataframe of median and 95\% CI for the pmf at each delay (in weeks)
}
\description{
Derive multipliers using MADPH methods but within this codebase, using
their original implementation
}
