% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_scores_fig.R
\name{get_plot_nowcasts_vs_data}
\alias{get_plot_nowcasts_vs_data}
\title{Get a plot illustrating nowcasts at certain dates}
\usage{
get_plot_nowcasts_vs_data(
  nowcasts,
  all_data,
  max_delay,
  pathogen_i,
  nowcast_dates_to_plot = NULL,
  facet = FALSE
)
}
\arguments{
\item{nowcast_dates_to_plot}{Vector of character strings of the dates you
wish to plot, default is \code{NULL} which will plot all of them}

\item{facet}{Boolean indicating whether or not to make separate facets
of each model}

\item{combined_nowcasts}{Dataframe of the combined quantiles across
horizons and nowcast dates}

\item{max_horizon}{Integer indicating the maximum horizon to plot.}

\item{age_group_to_plot}{Character string indicating which age group to
show in the plot, default is "00+" for all age groups.}
}
\value{
ggplot object
}
\description{
Get a plot illustrating nowcasts at certain dates
}
