% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_fig.R
\name{get_violin_plot_delay}
\alias{get_violin_plot_delay}
\title{Violin plot of delay by pathogen and age group}
\usage{
get_violin_plot_delay(weekly_data, season_to_plot = NULL, ylims = TRUE)
}
\arguments{
\item{weekly_data}{Dataframe of cases by reference date, report date, week,
and pathogen}

\item{season_to_plot}{Character string indicating what season to plot,
default is NULL which will include all dates in the data}
}
\value{
ggplot
}
\description{
Violin plot of delay by pathogen and age group
}
