% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_fig.R
\name{get_violin_plot_prop_visits}
\alias{get_violin_plot_prop_visits}
\title{Violin plot of proportion of visits}
\usage{
get_violin_plot_prop_visits(weekly_data, season_to_plot = NULL, weekly = TRUE)
}
\arguments{
\item{weekly_data}{Dataframe of cases by reference date, report date, week,
and pathogen}

\item{season_to_plot}{Character string indicating what season to plot,
default is NULL which will include all dates in the data}
}
\value{
ggplot
}
\description{
Violin plot of proportion of visits
}
