% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_scores_fig.R
\name{make_state_nowcast_comp_fig}
\alias{make_state_nowcast_comp_fig}
\title{Make state nowcast comparison figure}
\usage{
make_state_nowcast_comp_fig(
  nowcasts_vs_data,
  bar_chart_scores1,
  bar_chart_scores2,
  bar_chart_scores3,
  bar_chart_scores4,
  fig_file_name = NULL,
  fig_file_dir = file.path("output", "figs")
)
}
\arguments{
\item{nowcasts_vs_data}{plot A}

\item{bar_chart_scores1}{plot B}

\item{bar_chart_scores2}{plot C}

\item{bar_chart_scores3}{plot D}

\item{bar_chart_scores4}{plot E}

\item{fig_file_name}{}

\item{fig_file_dir}{}
}
\value{
patchwork fig
}
\description{
Make state nowcast comparison figure
}
