% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_style.R
\name{plot_components}
\alias{plot_components}
\title{Get plot components (colors and shapes)}
\usage{
plot_components()
}
\value{
a list of the model colors to be passed to \code{scale_fill_manual} and
\code{scale_color_manual}
}
\description{
Get plot components (colors and shapes)
}
